requirements.txt
codetiming==1.4.0
numpy==1.24.1
Pillow==9.4.0
pyTextColor==1.0.1
scipy==1.10.0
